frontend: extract search and setup controllers
