+++ src/components/Mask.jsx
@@ -1,145 +1,40 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { WifiOff, Loader2, Play, Route, PackageOpen, Shield, ShieldEllipsis, RotateCcw, Download, X } from 'lucide-react';
-import { open } from '@tauri-apps/plugin-dialog';
-import { invoke } from '@tauri-apps/api/core';
-import { listen } from '@tauri-apps/api/event';
+import { Loader2, Route, PackageOpen, Shield, RotateCcw, Download, X } from 'lucide-react';
+import { useRequiredModelDownload } from '../hooks/useRequiredModelDownload';
+import { useDepsSyncOverlay } from '../hooks/useDepsSyncOverlay';
+import { useVenvInstallController } from '../hooks/useVenvInstallController';
 
 
 export default function Mask({ backendStatus, pythonVersion, backendError, handleStartBackend, onRefreshPythonVersion, depsNeedUpdate, depsSyncing, onDepsSync, modelsNeedDownload, missingModels, onModelsDownloadComplete }) {
-
-  const [venvInstallStep, setVenvInstallStep] = React.useState(null);
-  const [pythonPath, setPythonPath] = React.useState('');
-  const [discoveredOptions, setDiscoveredOptions] = React.useState([]);
-  const [selectedVersions, setSelectedVersions] = React.useState([]);
-  const [versionErrorState, setVersionErrorState] = React.useState(null);
-  const [installing, setInstalling] = React.useState(false);
-  const [installError, setInstallError] = React.useState(null);
-
-  // Step 1: install button logs & helper (for auto-installing system Python)
-  const [installLogs, setInstallLogs] = React.useState([]);
-  const installLogRef = React.useRef(null);
-  const appendInstallLog = (msg) => {
-    setInstallLogs((prev) => [...prev, msg]);
-  };
-
-  // Global listener ref for backend install logs
-  const installListenerRef = React.useRef(null);
-
-  // Listen for 'install-log' events emitted from the Rust backend and route them
-  // to the appropriate log view (installer vs pip). This is registered once on
-  // mount and removed on unmount.
-  React.useEffect(() => {
-    let mounted = true;
-    (async () => {
-      try {
-        const unlisten = await listen('install-log', (event) => {
-          const payload = event?.payload || {};
-          const line = payload.line || JSON.stringify(payload);
-          const source = payload.source || 'installer';
-          // Route pip source to dependency logs, everything else goes to install logs
-          if (source === 'pip' || source === 'aria2') { 
-            appendDepsLog(line);
-          } else {
-            appendInstallLog(line);
-          }
-        });
-        installListenerRef.current = unlisten;
-      } catch (e) {
-        console.warn('Failed to register install-log listener', e);
-      }
-    })();
-
-    return () => {
-      if (installListenerRef.current) {
-        try {
-          installListenerRef.current();
-        } catch (e) {
-          console.warn('Failed to remove install-log listener', e);
-        }
-      }
-    };
-  }, []);
-
-  // Step 2: dependency installation (venv + pip install)
-  const [depsInstalling, setDepsInstalling] = React.useState(false);
-  const [depsInstallLog, setDepsInstallLog] = React.useState([]);
-  const [depsError, setDepsError] = React.useState(null);
-  const [depsInstallSuccess, setDepsInstallSuccess] = React.useState(false);
-  const depsLogRef = React.useRef(null);
-
-  // fetch discovered python installations asynchronously
-  const getDiscoveredPythonVersions = async () => {
-    const result = await invoke('check_python_status');
-    console.log(result);
-
-    // normalize result to an array of version strings
-    const versions = Array.isArray(result) ? result : [result].filter(Boolean);
-
-    return versions.map((pv, idx) => {
-      // if backend returned a full path (e.g. C:\\Program Files\\Python310\\python.exe), use it directly
-      const isPath = typeof pv === 'string' && (pv.includes('\\') || pv.includes('/') || String(pv).toLowerCase().endsWith('.exe'));
-      if (isPath) {
-        const path = pv;
-        const filename = path.split(/[\\\/]/).pop();
-        return {
-          id: `py${path.replace(/[^a-zA-Z0-9]/g, '') || idx}`,
-          display: `${filename} - ${path}`,
-          disabled: false,
-          path,
-        };
-      }
-
-      // otherwise assume pv is a version string like "3.10.11"
-      const verDigits = pv ? pv.replace(/\./g, '') : `${idx}`;
-      const path = `C:\\Python${verDigits}\\python.exe`;
-      return {
-        id: `py${verDigits}`,
-        display: `Python ${pv} - ${path}`,
-        disabled: false,
-        path,
-      };
-    });
-  };
-
-  React.useEffect(() => {
-    if (venvInstallStep === 1) {
-      let cancelled = false;
-      setVersionErrorState(null);
-      (async () => {
-        try {
-          setDiscoveredOptions([]); // clear while loading
-          const opts = await getDiscoveredPythonVersions();
-          if (!cancelled) setDiscoveredOptions(opts);
-        } catch (error) {
-          if (!cancelled) setVersionErrorState(error?.message || String(error));
-        }
-      })();
-      return () => { cancelled = true; };
-    }
-  }, [venvInstallStep]);
-
-  // Automatically run dependency installation when entering step 2
-  const appendDepsLog = (msg) => {
-    const ts = new Date().toLocaleTimeString();
-    setDepsInstallLog((prev) => [...prev, `[${ts}] ${msg}`]);
-  };
-
-  // Use a ref to track if installation has already started to prevent double execution
-  const installStartedRef = React.useRef(false);
-
-  // Store pythonPath in a ref so the effect always has access to the latest value
-  const pythonPathRef = React.useRef(pythonPath);
-  React.useEffect(() => {
-    pythonPathRef.current = pythonPath;
-  }, [pythonPath]);
-
-  // Reference to the actual input element so we can synchronously read the user's typed/selected path
-  const inputRef = React.useRef(null);
-  // Capture the python path that the user intended to use when they click "Next"
-  const [chosenPythonForInstall, setChosenPythonForInstall] = React.useState(null);
-  
   const { t } = useTranslation();
+  const {
+    venvInstallStep,
+    setVenvInstallStep,
+    pythonPath,
+    setPythonPath,
+    discoveredOptions,
+    selectedVersions,
+    versionErrorState,
+    installing,
+    installError,
+    installLogs,
+    installLogRef,
+    depsInstalling,
+    depsInstallLog,
+    depsError,
+    depsLogRef,
+    inputRef,
+    toggleVersion,
+    installPython,
+    openFileDialog,
+    beginDependencyInstall,
+    retryDependencyInstall,
+  } = useVenvInstallController({
+    onRefreshPythonVersion,
+    handleStartBackend,
+    t,
+  });
 
   // Debug helper: in dev mode show a dropdown to preview different masks
   const isDev = typeof import.meta !== 'undefined' && Boolean(import.meta.env && import.meta.env.DEV);
@@ -168,6 +63,34 @@ export default function Mask({ backendStatus, pythonVersion, backendError, handl
   const renderBackendStatus = debugOverrides.hasOwnProperty('backendStatus') ? debugOverrides.backendStatus : backendStatus;
   const renderPythonVersion = debugOverrides.hasOwnProperty('pythonVersion') ? debugOverrides.pythonVersion : pythonVersion;
   const renderVenvInstallStep = debugOverrides.hasOwnProperty('venvInstallStep') ? debugOverrides.venvInstallStep : venvInstallStep;
+  const {
+    modelDownloadLog,
+    modelDownloadError,
+    modelDownloading,
+    modelDownloadLogRef,
+    isClosedByUser,
+    setIsClosedByUser,
+    retryModelDownload,
+  } = useRequiredModelDownload({
+    modelsNeedDownload,
+    missingModels,
+    renderVenvInstallStep,
+    depsNeedUpdate,
+    onModelsDownloadComplete,
+    t,
+  });
+  const {
+    depsSyncLog,
+    depsSyncError,
+    depsSyncLogRef,
+    retryDepsSync,
+  } = useDepsSyncOverlay({
+    depsNeedUpdate,
+    pythonVersion,
+    renderVenvInstallStep,
+    depsSyncing,
+    onDepsSync,
+  });
 
   const renderDebugSelector = () => (
     <div className="absolute right-6 top-6 z-60 flex items-center gap-2">
@@ -188,439 +111,6 @@ export default function Mask({ backendStatus, pythonVersion, backendError, handl
       </select>
     </div>
   );
-
-  const autoPostInstallRef = React.useRef(false);
-
-  // ==================== Model download overlay state ====================
-  const [modelDownloadLog, setModelDownloadLog] = React.useState([]);
-  const [modelDownloadError, setModelDownloadError] = React.useState(null);
-  const [modelDownloading, setModelDownloading] = React.useState(false);
-  const modelDownloadLogRef = React.useRef(null);
-  const modelDownloadStartedRef = React.useRef(false);
-  const [isClosedByUser, setIsClosedByUser] = React.useState(false);
-  const [downloadProgressState, setDownloadProgressState] = React.useState({
-    keys: [],
-    currentIdx: 0,
-    currentFileProgress: {},
-  });
-
-  // Reset isClosedByUser when modelsNeedDownload becomes true
-  React.useEffect(() => {
-    if (modelsNeedDownload) {
-      setIsClosedByUser(false);
-    }
-  }, [modelsNeedDownload]);
-
-  // Memoize overall progress
-  const overallProgress = React.useMemo(() => {
-    const { keys, currentIdx, currentFileProgress } = downloadProgressState;
-    if (!keys || keys.length === 0) return 0;
-    const progressValues = Object.values(currentFileProgress);
-    const currentModelProgress = progressValues.length > 0
-      ? progressValues.reduce((sum, val) => sum + val, 0) / Math.max(progressValues.length, 5)
-      : 0;
-    const progress = (currentIdx * 100 + currentModelProgress) / keys.length;
-    return Math.max(0, Math.min(100, progress));
-  }, [downloadProgressState]);
-
-  // Dispatch custom event to let other components (like SearchBox) know about model download status
-  React.useEffect(() => {
-    window.dispatchEvent(
-      new CustomEvent('model-download-progress', {
-        detail: {
-          active: modelDownloading && isClosedByUser,
-          progress: Math.round(overallProgress),
-        },
-      })
-    );
-  }, [modelDownloading, isClosedByUser, overallProgress]);
-
-  // Auto-scroll model download log
-  React.useEffect(() => {
-    if (modelDownloadLogRef?.current) {
-      modelDownloadLogRef.current.scrollTop = modelDownloadLogRef.current.scrollHeight;
-    }
-  }, [modelDownloadLog]);
-
-  // Capture install-log events into model download log when downloading
-  React.useEffect(() => {
-    if (!modelDownloading) return;
-    let mounted = true;
-    let unlisten;
-    (async () => {
-      try {
-        unlisten = await listen('install-log', (event) => {
-          if (!mounted) return;
-          const payload = event?.payload || {};
-          const line = payload.line || JSON.stringify(payload);
-          const ts = new Date().toLocaleTimeString();
-          setModelDownloadLog((prev) => [...prev, `[${ts}] ${line}`]);
-
-          // Parse progress percentage from aria2 log line
-          if (payload.source === 'aria2' && payload.file) {
-            const match = line.match(/\((\d+)%\)/);
-            if (match) {
-              const percent = parseInt(match[1], 10);
-              setDownloadProgressState((prev) => {
-                const newProgress = { ...prev.currentFileProgress, [payload.file]: percent };
-                return {
-                  ...prev,
-                  currentFileProgress: newProgress,
-                };
-              });
-            }
-          }
-        });
-      } catch (e) {
-        console.warn('Failed to register model download log listener', e);
-      }
-    })();
-    return () => {
-      mounted = false;
-      if (unlisten) unlisten();
-    };
-  }, [modelDownloading]);
-
-  // Auto-start model download when needed
-  React.useEffect(() => {
-    if (!modelsNeedDownload || !missingModels) return;
-    if (modelDownloadStartedRef.current || modelDownloading) return;
-    // Don't show during first install or deps update
-    if (renderVenvInstallStep != null || depsNeedUpdate) return;
-
-    modelDownloadStartedRef.current = true;
-    setModelDownloading(true);
-    setModelDownloadLog([]);
-    setModelDownloadError(null);
-
-    (async () => {
-      try {
-        const config = await invoke('get_advanced_config');
-        const useOnnx = config?.use_onnx || false;
-
-        const keysToDownload = [];
-        if (missingModels['chinese-clip'] && !missingModels['chinese-clip'].complete) {
-          keysToDownload.push('chinese-clip');
-        }
-        if (missingModels['bge-small-zh'] && !missingModels['bge-small-zh'].complete) {
-          keysToDownload.push('bge-small-zh');
-        }
-        if (missingModels['minilm-l12'] && !missingModels['minilm-l12'].complete) {
-          keysToDownload.push('minilm-l12');
-        }
-
-        setDownloadProgressState({
-          keys: keysToDownload,
-          currentIdx: 0,
-          currentFileProgress: {},
-        });
-
-        for (let i = 0; i < keysToDownload.length; i++) {
-          const key = keysToDownload[i];
-          setDownloadProgressState((prev) => ({
-            ...prev,
-            currentIdx: i,
-            currentFileProgress: {},
-          }));
-
-          if (key === 'chinese-clip') {
-            setModelDownloadLog((prev) => [...prev, `[${new Date().toLocaleTimeString()}] ${t('mask.model_download.downloading_clip')}`]);
-            // This will fall back to default parameters inside Rust, which handles use_onnx
-            await invoke('download_model');
-          } else if (key === 'bge-small-zh') {
-            setModelDownloadLog((prev) => [...prev, `[${new Date().toLocaleTimeString()}] ${t('mask.model_download.downloading_bge')}`]);
-            if (useOnnx) {
-              await invoke('download_model', {
-                repo: 'Xenova/bge-small-zh-v1.5',
-                subdir: 'bge-small-zh-v1.5',
-                files: ['config.json', 'tokenizer.json', 'tokenizer_config.json', 'special_tokens_map.json', 'onnx/model_quantized.onnx'],
-                modelRuntime: 'onnx',
-              });
-            } else {
-              await invoke('download_model', {
-                repo: 'BAAI/bge-small-zh-v1.5',
-                subdir: 'bge-small-zh-v1.5',
-                files: ['config.json', 'pytorch_model.bin', 'tokenizer.json', 'tokenizer_config.json', 'vocab.txt', 'special_tokens_map.json'],
-              });
-            }
-          } else if (key === 'minilm-l12') {
-            setModelDownloadLog((prev) => [...prev, `[${new Date().toLocaleTimeString()}] ${t('mask.model_download.downloading_minilm')}`]);
-            if (useOnnx) {
-              await invoke('download_model', {
-                repo: 'Xenova/paraphrase-multilingual-MiniLM-L12-v2',
-                subdir: 'paraphrase-multilingual-MiniLM-L12-v2',
-                files: ['config.json', 'tokenizer.json', 'tokenizer_config.json', 'special_tokens_map.json', 'onnx/model_quantized.onnx'],
-                modelRuntime: 'onnx',
-              });
-            } else {
-              await invoke('download_model', {
-                repo: 'sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2',
-                subdir: 'paraphrase-multilingual-MiniLM-L12-v2',
-                files: ['config.json', 'pytorch_model.bin', 'tokenizer.json', 'tokenizer_config.json', 'special_tokens_map.json', 'sentencepiece.bpe.model'],
-              });
-            }
-          }
-        }
-
-        setModelDownloadLog((prev) => [...prev, `[${new Date().toLocaleTimeString()}] ${t('mask.model_download.complete')}`]);
-        if (onModelsDownloadComplete) onModelsDownloadComplete();
-      } catch (err) {
-        setModelDownloadError(err?.message || String(err));
-      } finally {
-        setModelDownloading(false);
-        modelDownloadStartedRef.current = false;
-      }
-    })();
-  }, [modelsNeedDownload, missingModels, modelDownloading, renderVenvInstallStep, depsNeedUpdate]);
-
-  // ==================== Deps update overlay state ====================
-  const [depsSyncLog, setDepsSyncLog] = React.useState([]);
-  const [depsSyncError, setDepsSyncError] = React.useState(null);
-  const [depsSyncStarted, setDepsSyncStarted] = React.useState(false);
-  const depsSyncLogRef = React.useRef(null);
-
-  // Auto-scroll deps sync log
-  React.useEffect(() => {
-    if (depsSyncLogRef?.current) {
-      depsSyncLogRef.current.scrollTop = depsSyncLogRef.current.scrollHeight;
-    }
-  }, [depsSyncLog]);
-
-  // Auto-start deps sync when overlay is shown
-  React.useEffect(() => {
-    if (!depsNeedUpdate || !pythonVersion || renderVenvInstallStep != null) return;
-    if (depsSyncStarted || depsSyncing) return;
-
-    setDepsSyncStarted(true);
-    setDepsSyncLog([]);
-    setDepsSyncError(null);
-
-    (async () => {
-      try {
-        await onDepsSync();
-        setDepsSyncStarted(false);
-      } catch (err) {
-        setDepsSyncError(err?.message || String(err));
-        setDepsSyncStarted(false);
-      }
-    })();
-  }, [depsNeedUpdate, pythonVersion, renderVenvInstallStep, depsSyncStarted, depsSyncing, onDepsSync]);
-
-  // Listen for install-log events to populate deps sync log
-  // (We reuse the existing listener — route 'pip' and 'installer' source lines to depsSyncLog when syncing)
-  const prevDepsNeedUpdate = React.useRef(false);
-  React.useEffect(() => {
-    // When depsNeedUpdate transitions from false to true, start capturing logs
-    if (depsNeedUpdate && !prevDepsNeedUpdate.current) {
-      setDepsSyncLog([]);
-      setDepsSyncError(null);
-    }
-    prevDepsNeedUpdate.current = depsNeedUpdate;
-  }, [depsNeedUpdate]);
-
-  // Capture install-log events into depsSyncLog when deps are syncing
-  React.useEffect(() => {
-    if (!depsNeedUpdate && !depsSyncing) return;
-    let mounted = true;
-    let unlisten;
-    (async () => {
-      try {
-        unlisten = await listen('install-log', (event) => {
-          if (!mounted) return;
-          const payload = event?.payload || {};
-          const line = payload.line || JSON.stringify(payload);
-          const ts = new Date().toLocaleTimeString();
-          setDepsSyncLog((prev) => [...prev, `[${ts}] ${line}`]);
-        });
-      } catch (e) {
-        console.warn('Failed to register deps sync log listener', e);
-      }
-    })();
-    return () => {
-      mounted = false;
-      if (unlisten) unlisten();
-    };
-  }, [depsNeedUpdate, depsSyncing]);
-
-  React.useEffect(() => {
-    if (venvInstallStep === 2) {
-      // Clear previous logs/errors so each run is fresh
-      setDepsInstallLog([]);
-      setDepsError(null);
-      setDepsInstallSuccess(false);
-      autoPostInstallRef.current = false;
-
-      // Prevent double execution (handles StrictMode/HMR by using a global flag)
-      if (window.__cp_install_started) {
-        appendDepsLog('安装已在进行中（忽略重复触发）');
-        return;
-      }
-      window.__cp_install_started = true;
-      installStartedRef.current = true;
-
-      setDepsInstalling(true);
-
-      // Prefer the value captured when the user clicked 'Next' (synchronously read from input),
-      // otherwise fall back to the input DOM value, then the ref value.
-      const currentPythonPath = (chosenPythonForInstall !== null && chosenPythonForInstall !== undefined)
-        ? chosenPythonForInstall
-        : (inputRef.current?.value || pythonPathRef.current);
-
-      // Escape backslashes by doubling them, since backend expects escaped backslashes
-      const processedPythonPath = currentPythonPath ? currentPythonPath.replace(/\\/g, '\\\\') : null;
-
-      // append initial log and perform an invoke to the backend
-      appendDepsLog(t('mask.venv.step2.log_start'));
-      appendDepsLog(t('mask.venv.step2.using_python', { path: processedPythonPath || t('mask.venv.step2.using_python_default') }));
-
-      (async () => {
-        try {
-          console.log('Calling install_python_venv invoke...', { python_path: processedPythonPath });
-          // try to call monitor IPC to ask it to install requirements
-          // note: the monitor may not be running yet; errors are caught and logged
-          // pass the python executable path chosen in the previous step (escaped)
-          const res = await invoke('install_python_venv', { python_path: processedPythonPath });
-          appendDepsLog(res);
-          const config = await invoke('get_advanced_config');
-          const useOnnx = config?.use_onnx || false;
-
-          // download model files (Chinese-CLIP)
-          appendDepsLog(t('mask.venv.step2.download_models'));
-          const modelRes = await invoke('download_model');
-          appendDepsLog(modelRes);
-
-          // download BGE classification model
-          appendDepsLog(t('mask.venv.step2.download_bge'));
-          const bgeRes = useOnnx
-            ? await invoke('download_model', {
-              repo: 'Xenova/bge-small-zh-v1.5',
-              subdir: 'bge-small-zh-v1.5',
-              files: ['config.json', 'tokenizer.json', 'tokenizer_config.json', 'special_tokens_map.json', 'onnx/model_quantized.onnx'],
-              modelRuntime: 'onnx',
-            })
-            : await invoke('download_model', {
-              repo: 'BAAI/bge-small-zh-v1.5',
-              subdir: 'bge-small-zh-v1.5',
-              files: ['config.json', 'pytorch_model.bin', 'tokenizer.json', 'tokenizer_config.json', 'vocab.txt', 'special_tokens_map.json'],
-            });
-          appendDepsLog(bgeRes);
-
-          // download MiniLM clustering model
-          appendDepsLog(t('mask.venv.step2.download_minilm'));
-          const minilmRes = useOnnx
-            ? await invoke('download_model', {
-              repo: 'Xenova/paraphrase-multilingual-MiniLM-L12-v2',
-              subdir: 'paraphrase-multilingual-MiniLM-L12-v2',
-              files: ['config.json', 'tokenizer.json', 'tokenizer_config.json', 'special_tokens_map.json', 'onnx/model_quantized.onnx'],
-              modelRuntime: 'onnx',
-            })
-            : await invoke('download_model', {
-              repo: 'sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2',
-              subdir: 'paraphrase-multilingual-MiniLM-L12-v2',
-              files: ['config.json', 'pytorch_model.bin', 'tokenizer.json', 'tokenizer_config.json', 'special_tokens_map.json', 'sentencepiece.bpe.model'],
-            });
-          appendDepsLog(minilmRes);
-          appendDepsLog(t('mask.venv.step2.deps_complete'));
-          setDepsInstallSuccess(true);
-        } catch (err) {
-          appendDepsLog(t('mask.venv.step2.deps_failed', { error: err?.message || String(err) }));
-          setDepsError(err?.message || String(err));
-          setDepsInstallSuccess(false);
-        } finally {
-          setDepsInstalling(false);
-          window.__cp_install_started = false; // allow retries
-          installStartedRef.current = false;
-        }
-      })();
-    } else {
-      // Reset the ref and global flag when leaving step 2
-      installStartedRef.current = false;
-      if (window.__cp_install_started) window.__cp_install_started = false;
-    }
-    // no cleanup required, we do not use intervals here
-  }, [venvInstallStep]);
-
-  React.useEffect(() => {
-    if (depsInstalling || depsError || !depsInstallSuccess) return;
-    if (autoPostInstallRef.current) return;
-    autoPostInstallRef.current = true;
-
-    (async () => {
-      try {
-        if (typeof onRefreshPythonVersion === 'function') {
-          await onRefreshPythonVersion();
-        }
-      } catch (e) {
-        console.warn('Failed to refresh Python version after install', e);
-      } finally {
-        setVenvInstallStep(null);
-        if (typeof handleStartBackend === 'function') {
-          handleStartBackend();
-        }
-      }
-    })();
-  }, [depsInstalling, depsError, depsInstallSuccess, handleStartBackend, onRefreshPythonVersion]);
-
-  React.useEffect(() => {
-    if (depsLogRef?.current) {
-      depsLogRef.current.scrollTop = depsLogRef.current.scrollHeight;
-    }
-  }, [depsInstallLog]);
-
-  // scroll install button log
-  React.useEffect(() => {
-    if (installLogRef?.current) {
-      installLogRef.current.scrollTop = installLogRef.current.scrollHeight;
-    }
-  }, [installLogs]);
-
-  const toggleVersion = (id) => {
-    const opt = discoveredOptions.find((o) => o.id === id);
-    if (!opt || opt.disabled) return;
-    if (selectedVersions.includes(id)) {
-      setSelectedVersions(selectedVersions.filter((s) => s !== id));
-    } else {
-      setSelectedVersions([...selectedVersions, id]);
-      // also set path to selected candidate for convenience
-      setPythonPath(opt.path || '');
-    }
-  };
-
-  const installPython = async () => {
-    setInstallError(null);
-    setInstalling(true);
-    setInstallLogs([]);
-    appendInstallLog(t('mask.venv.auto_install.log_start'));
-
-    try {
-
-      let result = await invoke('request_install_python');
-      appendInstallLog(t('mask.venv.auto_install.success'));
-
-      appendInstallLog(t('mask.venv.auto_install.finished'));
-      setVenvInstallStep(null);
-      let _ = await invoke('close_process');
-    } catch (err) {
-      setInstallError(t('mask.venv.install_failed', { error: err?.message || String(err) }));
-      appendInstallLog(t('mask.venv.install_failed', { error: err?.message || String(err) }));
-    } finally {
-      setInstalling(false);
-    }
-  };
-
-  const openFileDialog = async () => {
-    try {
-      const selected = await open({
-        multiple: false,
-        filters: [{ name: 'Executable', extensions: ['exe'] }]
-      });
-      if (!selected) return;
-      const chosen = Array.isArray(selected) ? selected[0] : selected;
-      setPythonPath(chosen);
-    } catch (err) {
-      console.error('Failed to open file dialog', err);
-    }
-  };
 
   // ==================== Deps update overlay ====================
   if (depsNeedUpdate && renderPythonVersion && renderVenvInstallStep == null) {
@@ -643,10 +133,7 @@ export default function Mask({ backendStatus, pythonVersion, backendError, handl
             <div className="mt-3 flex items-center gap-3">
               <span className="text-xs text-rose-400">{t('mask.deps_update.failed', { error: depsSyncError })}</span>
               <button
-                onClick={() => {
-                  setDepsSyncError(null);
-                  setDepsSyncStarted(false);
-                }}
+                onClick={retryDepsSync}
                 className="flex items-center gap-1 px-2 py-1 bg-blue-600 hover:bg-blue-700 text-white rounded text-xs transition-colors"
               >
                 <RotateCcw className="w-3 h-3" />
@@ -696,10 +183,7 @@ export default function Mask({ backendStatus, pythonVersion, backendError, handl
             <div className="mt-3 flex items-center gap-3">
               <span className="text-xs text-rose-400">{t('mask.model_download.failed', { error: modelDownloadError })}</span>
               <button
-                onClick={() => {
-                  setModelDownloadError(null);
-                  modelDownloadStartedRef.current = false;
-                }}
+                onClick={retryModelDownload}
                 className="flex items-center gap-1 px-2 py-1 bg-blue-600 hover:bg-blue-700 text-white rounded text-xs transition-colors"
               >
                 <RotateCcw className="w-3 h-3" />
@@ -807,7 +291,7 @@ export default function Mask({ backendStatus, pythonVersion, backendError, handl
 
         <div className="absolute right-6 bottom-6">
             <button
-            onClick={() => { setChosenPythonForInstall(inputRef.current?.value || pythonPath); setVenvInstallStep(2); }}
+            onClick={beginDependencyInstall}
             disabled={!canNext || installing}
             className="flex items-center gap-2 px-3 py-1.5 bg-blue-600 hover:bg-blue-700 text-white rounded text-sm transition-colors disabled:opacity-50"
           >
@@ -847,16 +331,7 @@ export default function Mask({ backendStatus, pythonVersion, backendError, handl
               <div className="mt-3 flex items-center gap-3">
               <span className="text-xs text-rose-400">{t('mask.venv.step2.install_failed')}</span>
               <button
-                onClick={() => {
-                  setDepsError(null);
-                  setDepsInstallLog([]);
-                  setDepsInstallSuccess(false);
-                  window.__cp_install_started = false;
-                  installStartedRef.current = false;
-                  // Re-trigger the installation effect by toggling step away and back
-                  setVenvInstallStep(null);
-                  setTimeout(() => setVenvInstallStep(2), 0);
-                }}
+                onClick={retryDependencyInstall}
                 className="flex items-center gap-1 px-2 py-1 bg-blue-600 hover:bg-blue-700 text-white rounded text-xs transition-colors"
               >
                 <RotateCcw className="w-3 h-3" />
